hf-space/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
feedparser>=6.0.11
yfinance>=0.2.40
supabase>=2.4.0
google-generativeai>=0.7.0
groq>=0.9.0
pandas>=2.2.0
pydantic>=2.7.0
python-dotenv>=1.0.0
